Isolate using namespaces
FIXME: PID namespace is blocking signals

+++ src/runner.cpp
@@ -9,41 +9,62 @@
 #include <sys/resource.h>
 #include <assert.h>
 #include <syscall.h>
+#include <sys/mount.h>
+#include <sys/mman.h>
+#include <sys/prctl.h>
 
-static pid_t sighandler_kill_pid;
+static pid_t static_child_pid = -1;
 void sigint_handler(int signo);
+extern char* our_stack;
 
 namespace jail {
 
+int Runner::clone_trampoline(void* arg) {
+    reinterpret_cast<jail::Runner*>(arg)->forkedChild();
+    jail::panic("returned to trampoline");
+}
+
 void Runner::run() {
-    pid_t pid = fork();
-    
+    char* stack = (char*)malloc(1024*1024);
+    int pid = clone(clone_trampoline, stack+(1024*1024), SIGCHLD|CLONE_NEWNS|CLONE_NEWNET|CLONE_NEWIPC|CLONE_NEWPID, this);
     if(pid < 0)
-        jail::panic("Runner fork failed");
-    
-    if(pid == 0) {
-        forkedChild();
-    } else {
-        std::cout<<"pid"<<pid<<'\n';
-        Runner::child_pid = pid;
-        forkedMonitor();
-    }
+        jail::panic("clone failed");
+
+    std::cout<<"pid"<<pid<<'\n';
+    Runner::child_pid = pid;
+    static_child_pid = pid;
+    forkedMonitor();
+    free(stack);
 }
 
 void Runner::forkedChild() {
+    prctl(PR_SET_PDEATHSIG, SIGKILL);
     int rc = ptrace(PTRACE_TRACEME,0,0,0);
     if(rc < 0)
         jail::panic("ptrace traceme failed");
-    
+
     const char* c_argv[Runner::exec_args.size()+2];
     c_argv[0] = Runner::exec_name.c_str();
     for(size_t i = 0; i < Runner::exec_args.size(); i++)
         c_argv[i+1] = Runner::exec_args[i].c_str();
     c_argv[2] = nullptr;
+    printf("c_argv %s.%s.%d\n", Runner::exec_name.c_str(), c_argv[1], Runner::child_pid);
+   
+    //assert(false);
+    assert(mount(nullptr, "/", nullptr, MS_PRIVATE|MS_REC, nullptr) == 0);
+    assert(mount("jail/", "jail/", NULL, MS_BIND, NULL) == 0);
+    assert(mount("jinp/", "jail/inp/", NULL, MS_BIND, NULL) == 0);
+    assert(mount("jinp/", "jail/inp/", NULL, MS_BIND|MS_REMOUNT|MS_RDONLY, NULL) == 0);
+    assert(mount(exec_name.c_str(), "jail/exe", NULL, MS_BIND, NULL) == 0);
+    assert(!chdir("jail"));
+    assert(!syscall(SYS_pivot_root, ".", "."));
+    assert(!umount2(".", MNT_DETACH));
+    assert(!setresgid(65534, 65534, 65534));
+    assert(!setresuid(65534, 65534, 65534));
     
     Runner::seccomp.attach();
     
-    execv(Runner::exec_name.c_str(), (char*const*) c_argv);
+    execv(/*exec_name.c_str()*/ "exe", (char*const*) c_argv);
     jail::panic("execv failed"); // this should not be reached
 }
 
@@ -80,11 +101,14 @@ void Runner::forkedMonitor() {
         jail::panic("wait failed");
     // this option creates additional event just before exit and allows to examine registers
     // change options only when stopped
-    rc = ptrace(PTRACE_SETOPTIONS, Runner::child_pid, 0, PTRACE_O_TRACEEXIT);
-    if(rc < 0)
-        jail::panic("prace exitkill failed");
+    // rc = ptrace(PTRACE_SETOPTIONS, Runner::child_pid, 0, PTRACE_O_TRACEEXIT);
+    // if(rc < 0)
+    //     jail::panic("prace setoptions failed");
+    std::cout<<"wait status exit:"<<WIFEXITED(wait_s)<<" sig:"<<WIFSIGNALED(wait_s)<<" stop:"<<WIFSTOPPED(wait_s)<<'\n';
+    std::cout<<"codes s"<<WSTOPSIG(wait_s)<<" t"<<WTERMSIG(wait_s)<<'\n';
     
-    rc = ptrace(PTRACE_CONT, Runner::child_pid, 0, 0);
+    rc = ptrace(PTRACE_CONT, Runner::child_pid, 0, WSTOPSIG(wait_s));
+    // debug sig 6 
     if(rc < 0)
         jail::panic("initial ptrace_cont failed");
     std::cout<<"resuming after initial stop\n";
@@ -114,8 +138,9 @@ void Runner::forkedMonitor() {
         long ret = ptrace(PTRACE_GETREGS, Runner::child_pid, 0, &uregs); // this is not available when exit status is set
         if(ret < 0)
             jail::panic("ptrace_getregs failed");
-        std::cerr<<"regs: rax="<<uregs.rax<<" orax="<<uregs.orig_rax<<" rip(pc)="<<uregs.rip<<'\n';
+        std::cout<<"regs: rax="<<uregs.rax<<" orax="<<uregs.orig_rax<<" rip(pc)="<<uregs.rip<<'\n';
         
+        std::cout<<"SENDSIG\n";
         ret = ptrace(PTRACE_CONT, Runner::child_pid, 0, WSTOPSIG(wait_s));
         if(ret < 0)
             jail::panic("ptrace_cont failed");
@@ -135,7 +160,6 @@ void Runner::forkedMonitor() {
 }
 
 void Runner::setIntHandler(bool enable) {
-    sighandler_kill_pid = Runner::child_pid;
     struct sigaction sigact;
     sigact.sa_handler = (enable ? sigint_handler : SIG_DFL);
     sigemptyset(&sigact.sa_mask);
@@ -144,13 +168,17 @@ void Runner::setIntHandler(bool enable) {
     sigaction(SIGTERM, &sigact, NULL);
 }
 
+void Runner::killChild() {
+    int pid = static_child_pid; // change to map if multiple runners
+    if(pid > 0 && kill(pid, 0) == 0)
+        kill(pid, SIGKILL);
+}
+
 };
 
 void sigint_handler(int signo) {
     (void) signo; // unused
     std::cout<<" Recived termination signal. Killing child process\n";
-    int pid = sighandler_kill_pid;
-    if(kill(pid, 0) == 0)
-        kill(pid, SIGKILL);
+    jail::Runner::killChild();
     jail::panic("Interrupted");
 }
